+++ app/src/androidTest/java/com/github/factotum_sdp/factotum/MainActivityTest.kt
@@ -1,7 +1,6 @@
-package com.github.schuetzcarl.bootcamp
+package com.github.factotum_sdp.factotum
 
 import androidx.test.espresso.Espresso.onView
-import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.action.ViewActions.*
 import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.intent.Intents
